ml/requirements.txt
PyYAML~=6.0.1
ultralytics~=8.2.39

annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
idna==3.7
mistralai==0.4.0
orjson==3.10.5
pydantic==2.7.4
pydantic_core==2.18.4
sniffio==1.3.1
typing_extensions==4.12.2
python-dotenv
pathlib
requests
fastapi==0.111.0
uvicorn==0.30.1
